modified route to auth route and add modified readme
